`starttime`: Get process start time from `/proc/self/status` if possible.
This more accurately determines when the process was started, since it is not
affected by the overhead of running linked library `init` code and
initializing `runsc`'s CLI code.

Additionally, when `runsc` execs itself, preserve this time across the exec.

Also move the code that determines the start time earlier in the file.

PiperOrigin-RevId: 741311782